made i unsigned long int

+++ hash_tables/5-hash_table_print.c
@@ -10,7 +10,7 @@
 void hash_table_print(const hash_table_t *ht)
 {
 	hash_node_t *list, *node;
-	int i, firstElmnt = 1;
+	unsigned long int i, firstElmnt = 1;
 
 	if (ht == NULL)
 		return;
